// src/app/notreduxs/notweapons/notweapons.service.ts
import { Injectable } from '@angular/core';
import { Actions, createEffect, ofType } from '@ngrx/effects';
import { cationsnotweaponseftches, cationsnotweaponseftcheserrors, sidpatchesnotweaponseftcheserrors } from './erducers';
import { map, switchMap } from 'rxjs';
import { notpodsnotaxios } from '../../notpodsnotaxios';
import { inotweaponsoutfalse, isidpatches, iweaponsouttrue } from '../interface';
import { AxiosResponse } from 'axios';
import { sidpatchesnotweaponslosteftches } from '../notweaponslosts/erducers';

@Injectable({
  providedIn: 'root'
})
export class NotweaponsService {

  constructor(
    private cation: Actions
  ) { }

  eftches = createEffect(() => {
    return this.cation.pipe(
      ofType(cationsnotweaponseftches),
      switchMap(ac => notpodsnotaxios.get('/gladiator-summa-bid-arma/' + ac.payload).then((req: AxiosResponse<iweaponsouttrue>) => {
        let sidpatches: isidpatches<inotweaponsoutfalse> = {
          type: sidpatchesnotweaponseftcheserrors,
          payload: {
            defenceds: req.data['obstructionum-arma'].defensio,
            attackeds: req.data['obstructionum-arma'].impetus,
            lowestnotbidsnotlibers: req.data['summa-bid-notliber'],
            lowestnotbidsnotfixums: req.data['summa-bid-notfixum']
          }
        };
        return sidpatches;
      }))
    )
  })
}
